requirements.txt
streamlit>=1.31.0
requests>=2.31.0
google-generativeai>=0.3.0
scikit-learn>=1.3.0
pandas>=2.0.0
python-dotenv>=1.0.0
plotly>=5.18.0
matplotlib>=3.8.0
